+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   url: string;
   ```
 * added `@IsInstance` decorator and `validator.isInstance(value, target)` methods.
+* changed `@IsNumber` decorator has been changed to `@IsNumber(allowNaN: boolean = false)`
 
 #### Fixes